Update reading time data structures and add HTML statistics extraction
- Modified DTOs and model to change EstimatedSeconds from int to float64 for better precision.
- Implemented ExtractHTMLtoStatistics function to calculate reading statistics from HTML content.
- Added bluemonday for HTML sanitization to ensure safe text processing.
- Updated utility functions to support new reading statistics calculations.

+++ internal/reading_time/dto.go
@@ -1,39 +1,39 @@
 package reading_time
 
 type CreateReadingTimeRequest struct {
-	Minutes          int    `json:"minutes" binding:"required"`
-	TextLength       int    `json:"text_length" binding:"required"`
-	EstimatedSeconds int    `json:"estimated_seconds" binding:"required"`
-	WordCount        int    `json:"word_count" binding:"required"`
-	Type             string `json:"type" binding:"required,oneof=Blog"`
+	Minutes          int     `json:"minutes" binding:"required"`
+	TextLength       int     `json:"text_length" binding:"required"`
+	EstimatedSeconds float64 `json:"estimated_seconds" binding:"required"`
+	WordCount        int     `json:"word_count" binding:"required"`
+	Type             string  `json:"type" binding:"required,oneof=Blog"`
 }
 
 type UpdateReadingTimeRequest struct {
-	Id               int    `json:"id" binding:"required"`
-	Minutes          int    `json:"minutes" binding:"required"`
-	TextLength       int    `json:"text_length" binding:"required"`
-	EstimatedSeconds int    `json:"estimated_seconds" binding:"required"`
-	WordCount        int    `json:"word_count" binding:"required"`
-	Type             string `json:"type" binding:"required,oneof=Blog"`
+	Id               int     `json:"id" binding:"required"`
+	Minutes          int     `json:"minutes" binding:"required"`
+	TextLength       int     `json:"text_length" binding:"required"`
+	EstimatedSeconds float64 `json:"estimated_seconds" binding:"required"`
+	WordCount        int     `json:"word_count" binding:"required"`
+	Type             string  `json:"type" binding:"required,oneof=Blog"`
 }
 
 type ReadingTimeResponse struct {
-	ID               int    `json:"id"`
-	Minutes          int    `json:"minutes"`
-	TextLength       int    `json:"text_length"`
-	EstimatedSeconds int    `json:"estimated_seconds"`
-	WordCount        int    `json:"word_count"`
-	Type             string `json:"type"`
-	CreatedAt        string `json:"created_at"`
+	ID               int     `json:"id"`
+	Minutes          int     `json:"minutes"`
+	TextLength       int     `json:"text_length"`
+	EstimatedSeconds float64 `json:"estimated_seconds"`
+	WordCount        int     `json:"word_count"`
+	Type             string  `json:"type"`
+	CreatedAt        string  `json:"created_at"`
 }
 
 type ReadingTimeUpdateResponse struct {
-	ID               int    `json:"id"`
-	Minutes          int    `json:"minutes"`
-	TextLength       int    `json:"text_length"`
-	EstimatedSeconds int    `json:"estimated_seconds"`
-	WordCount        int    `json:"word_count"`
-	Type             string `json:"type"`
+	ID               int     `json:"id"`
+	Minutes          int     `json:"minutes"`
+	TextLength       int     `json:"text_length"`
+	EstimatedSeconds float64 `json:"estimated_seconds"`
+	WordCount        int     `json:"word_count"`
+	Type             string  `json:"type"`
 }
 
 type ReadingTimeDeleteRequest struct {
